Fix logical errors in example
This fixes two inconsistencies between the input text and the results in the first activecode example. 

If the desired results is a list of all positive even numbers less than the input, then the results should not include zero or the input. 
Zero is unsigned, neither positive nor negative. The input number should only be included if the text was "less than or equal to".